backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
email-validator
pydantic-settings
python-jose[cryptography]
bcrypt
python-multipart
celery
redis
python-dotenv
qrcode[pil]
Pillow
cryptography
slowapi
limits
alembic
python-json-logger
httpx
